Fix competition mode marker offset transparency.
In competition mode, the offset for the competition mode markers was not being
taken away from the absolute codes. This has probably not been noticed because
most people use the relative marker offsets instead of absolute marker numbers
- the bug has existed from the start.

Could this be why I've heard anecdotal evidence of a few robots failing
to detect markers last year?

This patch set was actually tested, unlike the first.

Change-Id: I3ed0c7abcfee19c008ebd9abe93263a783798ecd