Add a script to strip debug symbols without creating a zip
On the main server, it would make much more sense to cache a copy
of the binaries pre-stripped rather than stripping them each time.
At the moment, 42% of the time creating a zip is spent just looking
at the binaries to see if they need stripping (I don't even have
the strip tooling on my machine, so this is without even doing the
stripping).