Add an optional argument for the team TLA
The equivalent script for the new kit is going to require this
argument. I'd like to be able to make and test the changes to
the IDE before that appears, which means having soemthing to
test against.

Making the argument optional avoids breaking anything else currently
relying on or using this script.

Change-Id: Ic45dc8b13fcc4f313fed2dea8882d55a67dfa7ba